Changing parser to perform two passes, which is what Xyce does in practice
The first passes gathers statistics on the netlist, number of devices, number
of nodes, types of devices, etc.  The second pass distributes them over processors.

This will allow miniXyce to be able to handle flattened netlist files from
Xyce for performance studies.  At this time, the circuits will need to be
linear, but that will be extended.