Additional work on the parser to make it more similar to Xyce
The comment line symbol has been changed to be the same as Xyce, and
'.' lines are ignored by the miniXyce parser.  The SINE source function
has been changed to SIN, to be the same as Xyce. Ideally, this would enable
the miniXyce simulator to give performance statistics using Xyce acceptable
netlists.

The test netlists and netlist generators have been updated to use the
changed syntax.