Commiting miniXyce 1.1
This new miniXyce simulator has a parser and device model package.
It performs a DCOP load and linear solve, then transient loads as
determined by the options file.  The linear device models are as
the same as in Xyce.  The nonlinear device models are surrogate models
developed by a random equation generator.

The linear algebra underlying the miniXyce 1.1 are truly distributed
memory, performing the essential communication for correctness in
computations.  This is similar to standard linear algebra implementations.